testing new db

+++ postgresser.py
@@ -8,22 +8,39 @@ class Postgresser:
 
     def add_user(self, user_id):
         with self.connection:
-            if self.cursor.execute(f"SELECT user_id FROM users WHERE user_id = {user_id}") != None:
+            print('connected to db')
+            self.cursor.execute(f"SELECT user_id FROM users WHERE user_id = {user_id}")
+            result = self.cursor.fetchone()
+            print(result)
+            if not result:
                 self.cursor.execute("INSERT INTO users (user_id, status) VALUES(%s, %s)", (user_id, True))
                 self.connection.commit()
-                return 'success'
+                print('added')
+                return 'success adding'
             else:
-                return 'existed_user'
+                print('already exists')
+                return 'existed user'
 
     def remove_user(self, user_id):
         with self.connection:
-            self.cursor.execute(f"DELETE FROM users WHERE user_id = {user_id}")
-            self.connection.commit()
+            print('connected to db')
+            self.cursor.execute(f"SELECT user_id FROM users WHERE user_id = {user_id}")
+            result = self.cursor.fetchone()
+            print(result)
+            if result:
+                self.cursor.execute(f"DELETE FROM users WHERE user_id = {user_id}")
+                self.connection.commit()
+                return 'success remove'
+            else:
+                print('user not exists')
+                return 'not existed user'
 
 
     def return_users(self):
         with self.connection:
+            print('connected to db')
             self.cursor.execute("SELECT user_id FROM users")
+            print('collected user ids')
             temp = self.cursor.fetchall()
             return [x for t in temp for x in t]
